Support storing user config without dedicated user table
This patch introduces a setting 'store_config_by_userid' which is true by default.
When set to false, the user config is stored in a table user_infos with the username as its primary key.
This allows storing user settings without requiring the usage of the qwc user management api, eg. by using the qwc-oidc-auth service.
The schema of the user_infos table omits the user_id column and otherwise depends on the user_info_fields config just as the default setup.